Revert "添加表面交互和材质基本关联"
This reverts commit 7c5373ee8825ea65495b94f51a85bff3b649d7a0.

+++ HyperionFrame/Include/Scenes/Shape.cpp
@@ -29,7 +29,7 @@ AABB Shape::GenerateAABB()
 
 UINT Shape::GetFaceCount()
 {
-	return (UINT)m_indices.size() / 3;
+	return m_indices.size() / 3;
 }
 
 HTriangle Shape::GetFace(UINT faceIndex)
